Decryption fallback moved to JCEEncryptorCmd plus unit test update
